requirements.txt
Flask
pybktree
psycopg2-binary
python-Levenshtein==0.12.0
